Update account stack navigation and fix recursive dependency
